Modificacoes gerais 26/05/2017

+++ 05-biblioteca/src/biblioteca/ItemBiblioteca.java
@@ -8,7 +8,7 @@ package biblioteca;
 
 public interface ItemBiblioteca {
 
-	public long getCodigoSequencial();
+	public int getCodigoSequencial();
 
 	public void setCodigoSequencial();
 }
